Version 0.1.0 - 2 degrees of Spatial Perspective
This app lets you visualize the size of your home as compared to the size
of your city, as well as visualize the size of your city compared to the
size of your country and the world.

Pressing Ctrl+S in a visualization window lets you save the graphic as a
PNG file.

The plan is to add a third spatial visualization, re: the earth's orbit vs
the sun, as well as population-related visualizations.