Using a true checkboard grid instead of a grid of HxV lines
This commit changes the grid-drawing algorithm such that
a true checkboard pattern is now achieved instead of the
former grid of horizontal & vertical lines.

Also, the on-screen grids now use a 2x2 pixel square as
the basic unit, to make the grid easier to perceive with
the naked eye. The saved PNGs still use a 1x1 pixel as the
base unit however, as you can always zoom in on those.